Update
Modified the speed of the avatar and tested, landing on 10% slower as a
good speed.

Changed the variable of the avatar speed to a float from an int.

Challenge room tiles now only appear every 7 tiles.

Fixed issue with Challenge room tiles where the doors spawned on top of
the boarders.

General Tile adjustments.

+++ Assets/Scripts/AvatarController.cs
@@ -8,7 +8,7 @@ public class AvatarController : MonoBehaviour {
     private Vector3 previous_avatar_position;   // The avatars position before it travelled min_distance 
     private string avatar_direction;            // The direction the avatar is moving
     private Vector3 avatar_vector_direction;    // The avatar vector direction
-    public int avatar_speed;                    // The speed at which the avatar moves
+    public float avatar_speed;                    // The speed at which the avatar moves
     public Vector2 default_avatar_position;     // The avatars starting position
     public TailMovement tail_movement_script;
     Dictionary<string, string> cw_movement = new Dictionary<string, string>();  // Defines the next direction for clockwise turn
